pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-lower-win"
version = "0.1.0"
description = "Paper-trading and profile research tools for low-probability Polymarket crypto up/down markets."
requires-python = ">=3.11"
dependencies = [
    "websocket-client>=1.8,<2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
